+++ lib/active_record_query_counter.rb
@@ -101,13 +101,13 @@ module ActiveRecordQueryCounter
       query_time_threshold = (counter.thresholds.query_time || -1)
       if query_time_threshold >= 0 && query_time >= query_time_threshold
         trace = backtrace
-        send_notification("query_time", start_time, notification_end_time, sql: sql, binds: binds, row_count: row_count, trace: trace, elapsed_time: elapsed_time)
+        send_notification("query_time", start_time, notification_end_time, sql: sql, binds: binds, row_count: row_count, trace: trace, elapsed_time: elapsed_time, gc_time: gc_time, cpu_time: cpu_time)
       end
 
       row_count_threshold = (counter.thresholds.row_count || -1)
       if row_count_threshold >= 0 && row_count >= row_count_threshold
         trace ||= backtrace
-        send_notification("row_count", start_time, notification_end_time, sql: sql, binds: binds, row_count: row_count, trace: trace, elapsed_time: elapsed_time)
+        send_notification("row_count", start_time, notification_end_time, sql: sql, binds: binds, row_count: row_count, trace: trace, elapsed_time: elapsed_time, gc_time: gc_time, cpu_time: cpu_time)
       end
     end